scopes-intake/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
email-validator
python-dotenv
jinja2
